+++ src/test/java/org/folio/rest/api/InstanceStorageTest.java
@@ -8,6 +8,7 @@ import static java.net.HttpURLConnection.HTTP_OK;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.awaitility.Awaitility.await;
+import static org.folio.HttpStatus.HTTP_CREATED;
 import static org.folio.rest.impl.StorageHelper.MAX_ENTITIES;
 import static org.folio.rest.support.HttpResponseMatchers.errorMessageContains;
 import static org.folio.rest.support.HttpResponseMatchers.errorParametersValueIs;
@@ -39,6 +40,7 @@ import static org.hamcrest.CoreMatchers.nullValue;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.anyOf;
 import static org.hamcrest.Matchers.both;
+import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.greaterThanOrEqualTo;
 import static org.hamcrest.Matchers.hasSize;
 import static org.hamcrest.Matchers.lessThanOrEqualTo;
@@ -1652,7 +1654,7 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
   public void testCrossTableQueries() {
     final String url = instancesStorageUrl("") + "?query=";
 
-    //////// create instance objects /////////////////////////////
+    //------ create instance objects -----------------------------
     JsonArray identifiers = new JsonArray();
     identifiers.add(identifier(UUID_ISBN, "9781473619777"));
     JsonArray contributors = new JsonArray();
@@ -1673,16 +1675,16 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
     createInstance(j1);
     createInstance(j2);
     createInstance(j3);
-    ////////////////////// done /////////////////////////////////////
+    //-------------------- done ------------------------------------
 
-    ///////// create location objects //////////////////////////////
+    //------- create location objects ------------------------------
     UUID loc1 = UUID.fromString("11111111-dee7-48eb-b03f-d02fdf0debd0");
     LocationUtility.createLocation(loc1, "location1", "IX/L1");
     UUID loc2 = UUID.fromString("99999999-dee7-48eb-b03f-d02fdf0debd0");
     LocationUtility.createLocation(loc2, "location2", "IX/L2");
-    /////////////////// done //////////////////////////////////////
+    //----------------- done --------------------------------------
 
-    /////////////////// create holdings records ///////////////////
+    //----------------- create holdings records -------------------
     JsonObject jho1 = new JsonObject();
     String holdings1Uuid = loc1.toString();
     jho1.put("id", UUID.randomUUID().toString());
@@ -1706,7 +1708,7 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
     createHoldings(jho1);
     createHoldings(jho2);
     createHoldings(jho3);
-    ////////////////////////done //////////////////////////////////////
+    //----------------------done --------------------------------------
 
     String url1 = url + urlEncode("title=Long Way to a Small Angry Planet* sortby title");
     String url2 = url + urlEncode("title=cql.allRecords=1 sortBy title");
@@ -1716,7 +1718,7 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
     String url5 = url + urlEncode(
       "title=cql.allRecords=1 and holdingsRecords.permanentLocationId=99999999-dee7-48eb-b03f-d02fdf0debd0 "
         + "sortby title");
-    //non existant - 0 results
+    //non existent - 0 results
     String url6 = url + urlEncode(
       "title=cql.allRecords=1 and holdingsRecords.permanentLocationId=abc* sortby holdingsRecords.permanentLocationId");
 
@@ -1926,7 +1928,7 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
     CompletableFuture<Response> createCompleted = new CompletableFuture<>();
     getClient().post(instancesStorageSyncUrl(""), instanceCollection, TENANT_ID,
       ResponseHandler.empty(createCompleted));
-    assertThat(createCompleted.get(30, SECONDS), statusCodeIs(HttpStatus.HTTP_CREATED));
+    assertThat(createCompleted.get(30, SECONDS), statusCodeIs(HTTP_CREATED));
 
     var createdInstances = instancesArray.stream()
       .map(JsonObject.class::cast)
@@ -1948,7 +1950,7 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
       assertFalse(instance.containsKey(METADATA_KEY)));
 
     final var createCompleted = createInstancesBatchSync(instanceCollection);
-    assertThat(createCompleted.get(30, SECONDS), statusCodeIs(HttpStatus.HTTP_CREATED));
+    assertThat(createCompleted.get(30, SECONDS), statusCodeIs(HTTP_CREATED));
 
     toList(instanceCollection.getJsonArray(INSTANCES_KEY)).forEach(this::assertMetadataExists);
   }
@@ -1958,7 +1960,7 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
     JsonObject instanceCollection = createRequestForMultipleInstances(3);
 
     final var createCompleted = createInstancesBatchSync(instanceCollection);
-    assertThat(createCompleted.get(30, SECONDS), statusCodeIs(HttpStatus.HTTP_CREATED));
+    assertThat(createCompleted.get(30, SECONDS), statusCodeIs(HTTP_CREATED));
 
     toList(instanceCollection.getJsonArray(INSTANCES_KEY)).forEach(item ->
       assertThat(getById(item.getString("id")).getJson().getString(STATUS_UPDATED_DATE_PROPERTY),
@@ -2045,7 +2047,7 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
     JsonObject instanceCollection = new JsonObject().put(INSTANCES_KEY, instancesArray);
 
     Response response = instancesStorageSyncClient.attemptToCreate("?upsert=true", instanceCollection);
-    assertThat(response, statusCodeIs(HttpStatus.HTTP_CREATED));
+    assertThat(response, statusCodeIs(HTTP_CREATED));
 
     assertExists(instancesArray.getJsonObject(0));
     assertExists(instancesArray.getJsonObject(1));
@@ -2063,6 +2065,16 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
 
     instanceMessageChecks.createdMessagePublished(
       getById(secondInstanceToCreate.getString("id")).getJson());
+  }
+
+  @Test
+  public void canPostSynchronousBatchWithoutIdsWithUpsertTrue() {
+    canPostSynchronousBatchWithoutIds("?upsert=true");
+  }
+
+  @Test
+  public void canPostSynchronousBatchWithoutIdsWithoutUpsert() {
+    canPostSynchronousBatchWithoutIds("");
   }
 
   @Test
@@ -2099,10 +2111,10 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
 
     // insert
     JsonArray instances = new JsonArray().add(uprooted(UUID.randomUUID())).add(temeraire(UUID.randomUUID()));
-    assertThat(postSynchronousBatchUnsafe(instances), statusCodeIs(HttpStatus.HTTP_CREATED));
+    assertThat(postSynchronousBatchUnsafe(instances), statusCodeIs(HTTP_CREATED));
     // unsafe update
     instances.getJsonObject(1).put("title", "surprise");
-    assertThat(postSynchronousBatchUnsafe(instances), statusCodeIs(HttpStatus.HTTP_CREATED));
+    assertThat(postSynchronousBatchUnsafe(instances), statusCodeIs(HTTP_CREATED));
     // safe update, env var should not influence the regular API
     instances.getJsonObject(1).put("title", "sunset");
     assertThat(postSynchronousBatch("?upsert=true", instances), statusCodeIs(409));
@@ -2359,7 +2371,7 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
     getClient().post(instancesStorageSyncUrl(""), instanceCollection, TENANT_ID,
       ResponseHandler.empty(createCompleted));
 
-    assertThat(createCompleted.get(30, SECONDS), statusCodeIs(HttpStatus.HTTP_CREATED));
+    assertThat(createCompleted.get(30, SECONDS), statusCodeIs(HTTP_CREATED));
 
     for (int i = 0; i < 3; i++) {
       final JsonObject instance = instancesArray.getJsonObject(i);
@@ -2771,6 +2783,25 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
     assertGetNotFound(instancesStorageUrl("/" + instancesArray.getJsonObject(2).getString("id")));
   }
 
+  private void canPostSynchronousBatchWithoutIds(String postBatchParams) {
+    var instancesArray = new JsonArray();
+    var instanceRecord = new JsonObject()
+      .put("source", "MARC")
+      .put("title", "Test-Instance")
+      .put("instanceTypeId", UUID_INSTANCE_TYPE.toString())
+      .put("_version", 1);
+    instancesArray.add(instanceRecord);
+    assertThat(postSynchronousBatch(postBatchParams, instancesArray), statusCodeIs(HTTP_CREATED));
+    var getCompleted = new CompletableFuture<Response>();
+    getClient().get(instancesStorageUrl("") + "?query=title=Test-Instance", TENANT_ID, json(getCompleted));
+    var response = TestBase.get(getCompleted);
+    assertThat(response, statusCodeIs(HTTP_OK));
+    var createdInstance = response.getJson().getJsonArray("instances").getJsonObject(0);
+    assertThat(createdInstance.getString("id"), notNullValue());
+    assertThat(createdInstance.getString("title"), equalTo("Test-Instance"));
+    instanceMessageChecks.createdMessagePublished(getById(createdInstance.getString("id")).getJson());
+  }
+
   private void createInstancesWithClassificationNumbers() throws InterruptedException,
     ExecutionException, TimeoutException {
 
@@ -3008,13 +3039,6 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
 
     natureOfContentIdsToRemoveAfterTest.add(natureOfContentTerm.getId());
     return response.getJson().mapTo(NatureOfContentTerm.class);
-  }
-
-  private void assertNotSuppressedFromDiscovery(JsonArray array) {
-    array.stream()
-      .map(JsonObject.class::cast)
-      .map(instance -> instance.getString("id"))
-      .forEach(this::assertNotSuppressedFromDiscovery);
   }
 
   private void assertNotSuppressedFromDiscovery(String id) {