+++ NEWS.md
@@ -54,7 +54,8 @@
 * Fix items not being updated on /holdings-storage/batch/synchronous ([MODINVSTOR-1368](https://folio-org.atlassian.net/browse/MODINVSTOR-1368))
 * Fix typo in Instance Notes type name. ([MODINVSTOR-1396](https://folio-org.atlassian.net/browse/MODINVSTOR-1396))
 * Fix item sorting by order as a number ([MODINVSTOR-1470](https://folio-org.atlassian.net/browse/MODINVSTOR-1470))
-* Fix POST /holdings-storage/batch/synchronous?upsert=true returns 500 error when holdings "id' is missing ([MODINVSTOR-1487](https://folio-org.atlassian.net/browse/MODINVSTOR-1487))
+* Fix POST /holdings-storage/batch/synchronous?upsert=true returns 500 error when holdings "id" is missing ([MODINVSTOR-1487](https://folio-org.atlassian.net/browse/MODINVSTOR-1487))
+* Fix POST /instance-storage/batch/synchronous instance creation with absent instance "id" ([MODINVSTOR-1488](https://folio-org.atlassian.net/browse/MODINVSTOR-1488))
 
 ### Tech Dept
 * Updates sample-data/instances JSON to include missing required fields that were introduced over the years.